CIV-9632 interest calc bulk (#85)
* interest calculation data  mapped

* point at ccd

* Update Jenkinsfile_CNP

* point at civil

* bump version of org.springframework.boot to 3.1.6

* bump version of org.springframework.boot to 3.1.6

* bump version of org.springframework.boot to 3.1.6

* bump version of org.springframework.boot to 3.1.6

* point at civil and CCD change

* remove civil and ccd pointing

+++ src/main/java/uk/gov/hmcts/reform/civil/requestbody/Interest.java
@@ -18,7 +18,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 public class Interest {
     @Schema(example = "7")
-    private Integer interestDailyAmount;
+    private Long interestDailyAmount;
     @NotNull(message = "Interest Owed Date should not be null")
     @JsonFormat(pattern = "yyyy-MM-dd")
     @Schema(example = "2023-07-25")
